+++ test/DotNetty.Buffers.Tests/AbstractByteBufferTests.cs
@@ -3372,7 +3372,7 @@ namespace DotNetty.Buffers.Tests
 
             public bool Process(byte value)
             {
-                this.Bytes[this.index--] = value;
+                this.Bytes[this.index--] = (byte)value;
                 return true;
             }
         }
@@ -3411,7 +3411,7 @@ namespace DotNetty.Buffers.Tests
 
             public bool Process(byte value)
             {
-                this.Bytes[this.index++] = value;
+                this.Bytes[this.index++] = (byte)value;
                 return true;
             }
         }
